refactor: separate calendar heatmap into multiple components
add heatmap grid component to make it so that calendar heatmap is separated into different components
