+++ src/components/home/activity-heatmap.tsx
@@ -2,7 +2,7 @@
 
 import useActivityHeatmap from "@/hooks/use-activity-heatmap";
 import { ErrorBoundary } from "react-error-boundary";
-import CalendarHeatmap from "../calendar-heatmap";
+import CalendarHeatmap from "../calendar/calendar-heatmap";
 import ErrorFallback from "../error-fallback";
 import LoadingSpinner from "../loading-spinner";
